# [Website] Cache docs in memory, speed up page loads during development