This only affects the website when it is hosted locally using express. The current version of the website is annoyingly sluggish, as the whole docs structure is parsed on each request.

In this PR, we store the result of extracting the Markdown sources in memory, significantly speeding up page loads. This does not improve first page load time, as expected, as we need to extract the docs when the server first starts up.

Any changes to the docs structure may require a server restart to take effect. This is rare enough that I don't think it is a blocker.

## Timing

Extracting the docs on each request takes around 8 seconds. Storing these in memory allows us to virtually remove this delay entirely. Example from a single data point:

```
# Prior to this PR
extract docs, on every request: 8302.748ms

# With caching
extract docs, first request: 8107.119ms
extract docs, loaded from cache: 0.135ms
```