@ericvicenti 2017-02-06
Mostly looking good. But I'm not sure why we need `memory-cache` in comparison to a normal variable