@hramos 2017-02-06
A normal variable would not persist across server requests.